Refactor SPARQL query  and add search query with label
